Added user registration functionality

+++ base/views.py
@@ -6,29 +6,46 @@ from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+
+from django.contrib.auth.forms import UserCreationForm
 
 def loginpage(request):
+  page = 'login'
+  if request.user.is_authenticated:
+    return redirect('home')
   if request.method == "POST":
     username = request.POST.get('username')
     password = request.POST.get('password')
-
-    try:
-      user = User.objects.get(username = username)
-    except:
-      messages.error(request, "User doesn't exists.")
     user = authenticate(request, username = username, password = password)
     if user is not None:
       login(request,user)
       return redirect('home')
     else:
       messages.error(request, "Enter a Valid username or Password.")
-  context = {}
+  context = {'page':page}
   return render(request,'base/login_register.html',context)
 
 
 def logoutUser(request):
   logout(request)
   return redirect('home')
+
+
+def registerPage(request):
+  form = UserCreationForm()
+  if request.method == 'POST':
+    form = UserCreationForm(request.POST)
+    if form.is_valid():
+      user = form.save(commit=False)
+      user.username = user.username.lower()
+      user.save()
+      login(request, user)
+      return redirect('home')
+    else:
+      messages.error(request,'AN error occured during registration')
+  return render(request,'base/login_register.html',{'form':form})
+
 
 def home(request):
   q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -63,6 +80,9 @@ def createRoom(request):
 def updateRoom(request,pk):
   room = Room.objects.get(id = pk)
   form = RoomForm(instance=room)
+#Owner can only edit 
+  if request.user != room.host:
+    return HttpResponse("You can't edit others POST")
   if(request.method == "POST"):
     form = RoomForm(request.POST,instance=room)
     if form.is_valid():
@@ -75,6 +95,9 @@ def updateRoom(request,pk):
 @login_required(login_url='login')
 def deleteRoom(request,pk):
   room = Room.objects.get(id = pk)
+  if request.user != room.host:
+    return HttpResponse("You can't edit others POST")
+  
   if request.method == "POST":
     room.delete()
     return redirect('home')
